Update glium, term and glob dependency
This fixes a dependency resolution problem when picking minimal
dependencies as allowed by SemVer. The underlying problem is with the
outdated `winapi = ^0.2`, which contains a chain to `advapi32-sys = *`
and finally `winapi = 0.0.1`, which is itself a faulty package (invalid
syntax). In the glium case, the dependency chain is much longer through
`glutin`, wayland, `dlib` and `libloading`.