Remove `Reader::scrach_buffer` field + resulting breaking API changes.
This commit is desirable, because it avoids copying of image data across
intermediate buffers.  Avoiding such copying seems important based on
the data gathered in
https://github.com/image-rs/image-png/discussions/416#discussioncomment-7436871
Removal of `Reader::scrach_buffer` was not included in the initial
series of commits used to test the copy avoidance approach on
performance, but it should have a similar effect.

Fixes https://github.com/image-rs/image-png/issues/417