Restore using `NoMoreImageData` errors.
https://github.com/image-rs/image-png/pull/496 has incorrectly started
to return `UnexpectedEof` in a situation when when there is no
`Read`-level EOF.  In particular, this may happen when an `IDAT` of
`fdAT` chunk has been fully decoded, decompressed, and flushed, but
still doesn't contain the expected number of image pixels.  In such a
situation the old `NoMoreImageData` error was appropriate.  In a sense,
this commit is a partial revert and/or refactoring of the earlier commit
https://github.com/image-rs/image-png/pull/496/commits/b5b0674a0cd50a1f3c1e09f699c04b168648a237

This commit fixes a timeout found locally by the `buf_independent`
fuzzer.  The repro case has been saved under
`fuzz/corpus/buf_independent/regressions`.