Stop depending on `std::simd` / using `#[feature(portable_simd)]`.
Results of local Chromium benchmarks have questioned whether usage of
`std::simd` is beneficial or harmful to the runtime performance of the
`png` crate.  This was discussed at
https://github.com/image-rs/image-png/issues/567 and ended with
disabling `std::simd` in Chromium's usage of the `png` crate - see
https://crrev.com/c/6090592.

The most recent field trial data confirms that the `png` crate performs
well when not using `std::simd` - see
https://github.com/image-rs/image-png/discussions/562#discussioncomment-13303307

Based on the above, this commit removes the optional dependency on
`std::simd`.  At the very least, this results in less code to maintain.