Special case complete chunk start
In the state machine we spend a lot of time in `update`. When a chunk
starts we parse out its length, typestr, and only then dispatch to the
actual data. This parser must be able to handle any type of partial data
however in the common case we have all of it available at once. We can
skip a significant amount of state transitions (that both rewrite memory
and do loops around a match of that state) by handling the common case
in a single code path; entering partial states only when it is necessary
due to the input buffer ending while we are in a partial state.

We previously already did this for `u32` values but now can do it for
the whole chunk header.