Stream filtered decompressed data directly
The argument to `update` is changed from `&mut Vec<u8>` to a private
wrapper into a buffer such that we have direct control over the data
layout in the output buffer. This avoids intermediate buffering that we
could not perform with a vector as the buffer can keep track of more
than its length.

The new argument can also represent dropping of image data inputs,
although doing so will also drop any zlib decompression errors.