Opportunistically skip zlib finalization
When we detect the end of stream during the input iteration we do not
have any remaining buffer that needs to be append in a separate loop. In
particular we can avoid writing the filled/available amount entirely
because those were already detected during the main iteration. This
saves us filtering a partial buffer, tearing the filter loop across two
calls (instruction cache), and touching attributes which we already know
to be correct at their final value.

This results in an unexpectedly large amount of speedup in particular
for images that fit into a single IDAT and preallocation.