Adaptive, bounded buffer pre-allocation
This avoids DOS by an image pretending to be very wide but never actually having
image data for its size. The buffer preallocation is now bounded. Additionally,
the buffer preallocation and tuning parameters are based on their rowline such
that we avoid shifting and allocating partial rows while not allocating more
than the image data is allowed to be (well, this is optimized for non-progressive
images.