Optimize around codegen of copy_within
The compiler does not know that we can copy anything behind
`self.filled` transparently without affecting the result. In my
interpretation, the open ended range copy can internally utilize that
uninit data can actually be read just fine and this has better
instruction selection with larger moves.

This `copy_within(self.prev..)` was what the previous implementation did
but now we no longer truncate the vector. So we need to be a little
careful not to overshoot the relevant range by too many bytes.