Delay shift back behavior if buffer half empty
As explained in the comment, this is sort of adaptive to the size of
decoding chunks as well as compression ratio. Shifting back is an
expensive operation that we only do for bookkeeping so we should aim to
minimize it, the constraint being that we are not to starve the more
expensive decompression. The condition introduced here implies we're
probably not doing the latter.