Rework subbit adam7 iteration
This avoids overflows while calculating the read bit position in
scanlines of Adam7 interlacing. (Such an overflow is relevant for 32-bit
targets). This is part of a commit series, the other side of the coin is
not yet fixed.

The new method also vectorizes much better. Calling `filter` on an
iterator breaks a bunch of potential optimizations such as predicting
its precise length and thus loop conditions when filling. We avoid this
by instead calling a differently monomorphized subpixel iteration for
each possible bits-per-pixel.