Avoid overflowing multiplications in Adam7 writes
We now use two different iteration strategies. When we need bit
positions we calculate them precisely with a separate field for bytes
and the shift. When we have multi-byte samples (the common and
performance sensitive case of Rgb8 etc.) we fully any roundtrip through
bit-byte-bit calculations that would only put three 0-bits at the end of
each number to strip them again.