Regression test adam7 overflows
The original sketch was: <https://github.com/image-rs/image-png/pull/600/commits/4f1963bf80f37ac362cf0133b70324fc741cb30c>

In contrast to this we assert that the last bytes are filled and ensure
that the fill pattern and non-filled pattern of bits are
deterministically different, i.e. we use bits 0 and 1 for raw and filled
bytes in the buffer with no ambiguity.

Co-Authored-By: thirumurugan-git <thiru123arul@outlook.com>