Export splatted expansion functions
This is guided by Chromium's description of use here:
<https://github.com/image-rs/image-png/pull/609#issuecomment-3014429093>

We want to make sure that similar use cases pick up any performance work
we do on these functions as previous experience with the non-splatted
alternatives has shown there to be much potential.

Also removes the option from the decoder as there is no entry point
wherein a partial reconstruction of an image under the use of interlaced
expansion is every publicly exposed. The only application of these
functions occurs within `read_frame` that is supposed to return the
final result, i.e. the choice of algorithm is irrelevant up to
performance and we should choose for the user.